+++ src/bunker_monte_carlo.c
@@ -803,8 +803,8 @@ static RET_VAL _CalculatePropensities( BUNKER_MONTE_CARLO_RECORD *rec ) {
 
 static RET_VAL _CalculatePropensity( BUNKER_MONTE_CARLO_RECORD *rec, REACTION *reaction ) {
     RET_VAL ret = SUCCESS;
-    long stoichiometry = 0;
-    long amount = 0;
+    double stoichiometry = 0;
+    double amount = 0;
     double propensity = 0.0;
     double time = rec->time;
     SPECIES *species = NULL;
@@ -816,9 +816,9 @@ static RET_VAL _CalculatePropensity( BUNKER_MONTE_CARLO_RECORD *rec, REACTION *r
     edges = GetReactantEdges( (IR_NODE*)reaction );
     ResetCurrentElement( edges );
     while( ( edge = GetNextEdge( edges ) ) != NULL ) {
-        stoichiometry = (long)GetStoichiometryInIREdge( edge );
+        stoichiometry = GetStoichiometryInIREdge( edge );
         species = GetSpeciesInIREdge( edge );
-        amount = (long)GetAmountInSpeciesNode( species );
+        amount = GetAmountInSpeciesNode( species );
         if( amount < stoichiometry ) {
             if( IS_FAILED( ( ret = SetReactionRate( reaction, 0.0 ) ) ) ) {
                 return ret;
@@ -834,7 +834,7 @@ static RET_VAL _CalculatePropensity( BUNKER_MONTE_CARLO_RECORD *rec, REACTION *r
     edges = GetModifierEdges( (IR_NODE*)reaction );
     ResetCurrentElement( edges );
     while( ( edge = GetNextEdge( edges ) ) != NULL ) {
-        stoichiometry = (long)GetStoichiometryInIREdge( edge );
+        stoichiometry = GetStoichiometryInIREdge( edge );
         species = GetSpeciesInIREdge( edge );
         amount = GetAmountInSpeciesNode( species );
         if( amount < stoichiometry ) {
@@ -1106,7 +1106,7 @@ static void ExecuteAssignments( BUNKER_MONTE_CARLO_RECORD *rec ) {
 
 static RET_VAL _UpdateSpeciesValues( BUNKER_MONTE_CARLO_RECORD *rec ) {
     RET_VAL ret = SUCCESS;
-    long stoichiometry = 0;
+    double stoichiometry = 0;
     double amount = 0;
     SPECIES *species = NULL;
     IR_EDGE *edge = NULL;
@@ -1163,10 +1163,10 @@ static RET_VAL _UpdateSpeciesValues( BUNKER_MONTE_CARLO_RECORD *rec ) {
       edges = GetReactantEdges( (IR_NODE*)reaction );
       ResetCurrentElement( edges );
       while( ( edge = GetNextEdge( edges ) ) != NULL ) {
-        stoichiometry = (long)GetStoichiometryInIREdge( edge );
+        stoichiometry = GetStoichiometryInIREdge( edge );
         species = GetSpeciesInIREdge( edge );
 	if (HasBoundaryConditionInSpeciesNode(species)) continue;
-        amount = GetAmountInSpeciesNode( species ) - (double)stoichiometry;
+        amount = GetAmountInSpeciesNode( species ) - stoichiometry;
         TRACE_3( "the amount of %s decreases from %g to %g",
 		 GetCharArrayOfString( GetSpeciesNodeName( species ) ),
 		 GetAmountInSpeciesNode( species ),
@@ -1179,10 +1179,10 @@ static RET_VAL _UpdateSpeciesValues( BUNKER_MONTE_CARLO_RECORD *rec ) {
       edges = GetProductEdges( (IR_NODE*)reaction );
       ResetCurrentElement( edges );
       while( ( edge = GetNextEdge( edges ) ) != NULL ) {
-        stoichiometry = (long)GetStoichiometryInIREdge( edge );
+        stoichiometry = GetStoichiometryInIREdge( edge );
         species = GetSpeciesInIREdge( edge );
 	if (HasBoundaryConditionInSpeciesNode(species)) continue;
-        amount = GetAmountInSpeciesNode( species ) + (double)stoichiometry;
+        amount = GetAmountInSpeciesNode( species ) + stoichiometry;
         TRACE_3( "the amount of %s increases from %g to %g",
 		 GetCharArrayOfString( GetSpeciesNodeName( species ) ),
 		 GetAmountInSpeciesNode( species ),
